Changed from asyncio to multiprocessing
Asyncio wasn't working, the flask server wouldn't start untill the
metrics collection had finished. This is because asyncio operations are
thread blocking if they are CPU bound, which we are. Multiprocessing
instead allows us to run the metrics collection and flask server in
entirely different system processes that each have their own thread.