Bugfix for metrics collection
We've been using the psutil `oneshot` context manager method to read all
of the available metrics from a Process. This did not seem to work at
times and we would get processes who's metrics just had zeros. This
update instead uses the `to_dict` method so that the values are truely
frozen, and we can safely build ORM objects out of them.